backend/requirements.txt
Django==4.0.1
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-rest-enumfield==0.2.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
